+++ changelog.md
@@ -1,3 +1,2 @@
-- fixed an issue on level close possibly causing a leak
-- fixed possible issue when trying to connect to a vanilla server?
-- fixed wrong version number
+- fixed a bug with config screen icons
+- added custom config screen on forge